исправлена ошибка обновления цены

+++ lib/actions/shopOptpricePluginFrontendOptprice.controller.php
@@ -9,7 +9,7 @@ class shopOptpricePluginFrontendOptpriceController extends waJsonController {
     public function execute() {
         $product_id = waRequest::post('product_id');
         $sku_id = waRequest::post('sku_id');
-        $opt_price = shopOptpricePlugin::getOptPrice($product_id, $sku_id);
+        $opt_price = (float)shopOptpricePlugin::getOptPrice($product_id, $sku_id);
         if($opt_price) {
             $opt_price = shop_currency_html($opt_price);
         }
